# Lower memory usage from compiling and running binaries