--- a/cmd/editor/ide/tab.go
+++ b/cmd/editor/ide/tab.go
@@ -67,6 +67,8 @@ func (t *Tab) watchTitles(ctx context.Context, tabber Tabber) {
 		case title, ok := <-titles:
 			if ok {
 				t.title.SetInnerText(title)
+			} else {
+				t.Close()
 			}
 		}
 	}