--- a/internal/blob/blob.go
+++ b/internal/blob/blob.go
@@ -4,6 +4,7 @@ type Blob interface {
 	jsExtensions
 	Bytes() []byte
 	Len() int
+	View(start, end int64) (Blob, error)
 	Slice(start, end int64) (Blob, error)
 	Set(w Blob, off int64) (n int, err error)
 	Grow(off int64) error