--- a/internal/fs/wasm_cache.go
+++ b/internal/fs/wasm_cache.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/johnstarich/go-wasm/internal/blob"
 	"github.com/johnstarich/go-wasm/internal/fsutil"
-	"github.com/johnstarich/go-wasm/internal/indexeddb"
 	"github.com/johnstarich/go-wasm/internal/interop"
 	"github.com/johnstarich/go-wasm/internal/promise"
 	"github.com/johnstarich/go-wasm/log"
@@ -27,14 +26,12 @@ const (
 
 type wasmCacheFs struct {
 	rootFs
-	db       *indexeddb.DB
 	jsPaths  interop.StringCache
 	memCache map[string]js.Value
-	idbFound map[string]bool
 }
 
 func init() {
-	go initWasmCache()
+	initWasmCache()
 }
 
 func initWasmCache() {
@@ -47,16 +44,10 @@ func initWasmCache() {
 }
 
 func newWasmCacheFs(underlying rootFs) (*wasmCacheFs, error) {
-	db, err := indexeddb.New(wasmCacheDB, wasmCacheVersion, func(db *indexeddb.DB, oldVersion, newVersion int) error {
-		_, err := db.CreateObjectStore(wasmCacheStore, indexeddb.ObjectStoreOptions{})
-		return err
-	})
 	return &wasmCacheFs{
 		rootFs:   underlying,
-		db:       db,
 		memCache: make(map[string]js.Value),
-		idbFound: make(map[string]bool),
-	}, err
+	}, nil
 }
 
 func (w *wasmCacheFs) readFile(path string) (blob.Blob, error) {
@@ -87,30 +78,15 @@ func (w *wasmCacheFs) WasmInstance(path string, importObject js.Value) (js.Value
 		log.Debug("memCache hit: ", path)
 	} else {
 		log.Debug("memCache miss: ", path)
-		dbFound, dbChecked := w.idbFound[path]
-		if dbFound || !dbChecked {
-			moduleBlob, err := w.getModuleBytes(path)
-			if err != nil {
-				log.Debug("Failed to look up module at path: ", path)
-				return js.Value{}, err
-			}
-			if moduleBlob != nil {
-				module = moduleBlob.JSValue()
-			}
-		}
-	}
-
-	dbCacheHit := module.Truthy()
-	if dbCacheHit {
-		log.Debug("dbCache hit: ", path, module.Length())
-	} else {
-		log.Debug("dbCache miss: ", path)
 		moduleBlob, err := w.readFile(path)
 		if err != nil {
 			log.Debug("reading file failed: ", path)
 			return js.Value{}, err
 		}
 		module = moduleBlob.JSValue()
+		if !module.Truthy() {
+			log.Debug("fs miss: ", path, module.Length())
+		}
 	}
 
 	instantiatePromise := promise.From(jsWasm.Call("instantiate", module, importObject))
@@ -128,91 +104,13 @@ func (w *wasmCacheFs) WasmInstance(path string, importObject js.Value) (js.Value
 	}
 
 	w.memCache[path] = result.Get("module") // save compiled module for reuse
-	instance := result.Get("instance")
-	if !dbCacheHit {
-		// save module to IDB
-		return instance, w.setModule(path, module)
-	}
-	return instance, nil
+	return result.Get("instance"), nil
 }
 
 func (w *wasmCacheFs) dropModuleCache(path string) error {
 	path = fsutil.NormalizePath(path)
-
-	_, memCacheHit := w.memCache[path]
-	if memCacheHit {
-		delete(w.memCache, path)
-	}
-
-	dbCacheHit := w.idbFound[path]
-	if dbCacheHit {
-		err := w.deleteModule(path)
-		if err != nil {
-			return err
-		}
-		delete(w.idbFound, path)
-	}
+	delete(w.memCache, path)
 	return nil
-}
-
-func (w *wasmCacheFs) deleteModule(path string) error {
-	txn, err := w.db.Transaction(indexeddb.TransactionReadWrite, wasmCacheStore)
-	if err != nil {
-		return err
-	}
-	store, err := txn.ObjectStore(wasmCacheStore)
-	if err != nil {
-		return err
-	}
-	req, err := store.Delete(w.jsPaths.Value(path))
-	if err != nil {
-		return err
-	}
-	_, err = req.Await()
-	return err
-}
-
-func (w *wasmCacheFs) setModule(path string, module js.Value) error {
-	path = fsutil.NormalizePath(path)
-	txn, err := w.db.Transaction(indexeddb.TransactionReadWrite, wasmCacheStore)
-	if err != nil {
-		return err
-	}
-	store, err := txn.ObjectStore(wasmCacheStore)
-	if err != nil {
-		return err
-	}
-	req, err := store.Put(w.jsPaths.Value(path), module.JSValue())
-	if err != nil {
-		return err
-	}
-	_, err = req.Await()
-	if err == nil {
-		w.idbFound[path] = true
-	}
-	return err
-}
-
-func (w *wasmCacheFs) getModuleBytes(path string) (blob.Blob, error) {
-	path = fsutil.NormalizePath(path)
-	txn, err := w.db.Transaction(indexeddb.TransactionReadOnly, wasmCacheStore)
-	if err != nil {
-		return nil, err
-	}
-	store, err := txn.ObjectStore(wasmCacheStore)
-	if err != nil {
-		return nil, err
-	}
-	req, err := store.Get(w.jsPaths.Value(path))
-	if err != nil {
-		return nil, err
-	}
-	jsBytes, err := req.Await()
-	if err != nil || jsBytes.IsUndefined() {
-		return nil, err
-	}
-	w.idbFound[path] = true
-	return blob.NewFromJS(jsBytes)
 }
 
 func (w *wasmCacheFs) Create(name string) (afero.File, error) {
@@ -260,26 +158,4 @@ func (w *wasmCacheFs) Rename(oldname, newname string) error {
 
 func (w *wasmCacheFs) Chtimes(name string, atime time.Time, mtime time.Time) error {
 	return w.rootFs.Chtimes(name, atime, mtime)
-}
-
-func (w *wasmCacheFs) DestroyMount(path string) error {
-	txn, err := w.db.Transaction(indexeddb.TransactionReadWrite, wasmCacheStore)
-	if err != nil {
-		return err
-	}
-	store, err := txn.ObjectStore(wasmCacheStore)
-	if err != nil {
-		return err
-	}
-	req, err := store.Clear()
-	if err != nil {
-		return err
-	}
-	_, err = req.Await()
-	if err != nil {
-		return err
-	}
-	w.idbFound = make(map[string]bool)
-	w.memCache = make(map[string]js.Value)
-	return w.rootFs.DestroyMount(path)
 }