--- a/internal/rwonly/read_only_blob.go
+++ b/internal/rwonly/read_only_blob.go
@@ -1,5 +1,3 @@
-// +build js
-
 package rwonly
 
 import (