… cmd/editor/ide/settings.html
… internal/fs/write.go
… internal/interop/funcs.go
… internal/interop/profile.go
… internal/mountfs/fs.go
… internal/process/process_js.go
… internal/process/wasm.go
… internal/rwonly/write_only_blob.go
… internal/storer/file.go
… internal/storer/fs.go
… internal/terminal/term.go
… main.go
… server/public/index.html
… server/src/GoWASM.js