Lowers memory usage for running binaries like `go`, `compile`, and user programs.

Memory decreased by up to 56% in the best case and 47% in the worst case. The heap footprint should now remain below 250 MB in the worst case, down from 540 MB. On a second boot (not installing the Go toolchain), memory usage is reduced even further to below 180 MB. Execution times shortened a bit, but probably within margin of error.

See below for single samples of performance changes. These are super rough, so take them with a grain of salt.

Master branch:
* First boot:
    - Init = 59.04s; 297 MB heap
    - Compile = 179.18s
    - Run = 0.20s; 540 MB heap
* Second boot (after reload):
    - Init = 0.85s; 224 MB heap
    - Compile = 27.95s
    - Run = 0.12s; 394 MB heap

This PR:
* First boot:
	- Init = 64.35s; 156 MB heap `-47%`
	- Compile = 164.41s
	- Run = 0.13s; 245 MB heap `-55%`
* Second boot:
	- Init = 0.94s; 104 MB heap
	- Compile = 25.06s
	- Run = 0.10s; 175 MB heap `-56%`

I'd summarize the commits, but they were enough changes it'd probably be best to read through the commit messages themselves.